small LightingValue casting fixup. Shouldn't affect anything.

+++ Vixen.System/Data/Value/LightingValue.cs
@@ -16,7 +16,7 @@ namespace Vixen.Data.Value {
 
 		public Color GetIntensityAffectedColor() {
 			// as this is a lighting value, the lower the intensity (brightness), the more transparent it should be.
-			return Color.FromArgb((byte)(Intensity * byte.MaxValue), (int)(Color.R * Intensity), (int)(Color.G * Intensity), (int)(Color.B * Intensity));
+			return Color.FromArgb((int)(Intensity * byte.MaxValue), (int)(Color.R * Intensity), (int)(Color.G * Intensity), (int)(Color.B * Intensity));
 		}
 	}
 }
